refacto: moving `merge_models` into module `merge_models.py`
Because `pylint` currently fails to handle some operations on `pd.DataFrame` being performed in function `merge_models`,
and because disabling pylint is only possible module-wide, moving function merge_models into its own module `predictsignauxfaibles.merge_models`

Moreover:
- moving other functions from `predictsignauxfaibles.make_list` to `predictsignauxfaibles.utils`
- creating doctype for function `utils.assign_flag`
- correcting for function names imported from `utils` in `notebooks/04-merge_models_into_list.ipynb`
